refactor(agent-loop): split typed execution state
