+++ crates/loop/ironclaw_agent_loop/src/state/tests.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod checkpoint_wire;
+mod lifecycle;
